+++ packages/components/src/common/propsValues/scroll.ts
@@ -2,3 +2,6 @@ export enum Scroll {
   CONTENT = 'content',
   VIEWPORT = 'viewport',
 }
+
+export type ScrollContent = 'content';
+export type ScrollViewport = 'viewport';