+++ packages/docs/liveEditorCode/actionButton.code.js
@@ -1,0 +1,1 @@
+<ActionButton onClick={() => alert('clicked')} text="Manage"/>;