+++ packages/docs/utils/sections.ts
@@ -70,7 +70,7 @@ const sections: MainSection[] = [
   {
     title: 'Styles',
     directory: 'styles',
-    badge: { type: 'new', expiryDate: '2022-01-31' },
+    badge: { type: 'new', expiryDate: '2022-02-19' },
     children: [
       {
         title: 'Tokens',
@@ -85,7 +85,13 @@ const sections: MainSection[] = [
       {
         title: 'Addons',
         directory: 'styles/addons',
-        fileOrder: ['BackgroundUtilities', 'DisplayUtilities', 'RtlLayout', 'SpacingUtilities'],
+        fileOrder: [
+          'BackgroundUtilities',
+          'DisplayUtilities',
+          'DisabledState',
+          'RtlLayout',
+          'SpacingUtilities',
+        ],
       },
       {
         title: 'Extras',